refactor state functions
move collect_state and collect_compare_diffs from saas_file_owners to openshift_saas_deploy_change_tester. saas_file_owners will be removed in the near future so moving those functions that will survive makes sense.

at the same time those functions have been prepared for easier unittesting.
